+++ app/Models/Admin.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Notifications\Notifiable;
@@ -13,7 +13,7 @@ class Admin extends Authenticatable
     protected $fillable = [
         'name', 'email', 'password',
     ];
- 
+
     protected $hidden = [
         'password', 'remember_token',
     ];
